Akaros doesn't interpret "" as 127.0.0.1
Not sure if's supposed to or not, but as it stands, when you dial an ip
address of "ip:port" with the ip address part set to "", akaros
dosn't know what to do with this.  More precisely, cs doesn't know what
to do with this.  I imagine linux must translate this unde the hood, but
I haven't looked to deeply into how.  I need to look closer at whether
or not it's considered portable to not speficy the ip.